# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TokenOS Runtime",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/runs": {
      "post": {
        "summary": "Create Run",
        "operationId": "create_run_v1_runs_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSpec-Input"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "summary": "Read Run",
        "operationId": "read_run_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/reservations": {
      "post": {
        "summary": "Reserve",
        "operationId": "reserve_v1_runs__run_id__reservations_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReservationSpec"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/reservations/{reservation_id}/dispatch": {
      "post": {
        "summary": "Dispatch",
        "operationId": "dispatch_v1_runs__run_id__reservations__reservation_id__dispatch_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/reservations/{reservation_id}/settle": {
      "post": {
        "summary": "Settle",
        "operationId": "settle_v1_runs__run_id__reservations__reservation_id__settle_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/reservations/{reservation_id}/unknown": {
      "post": {
        "summary": "Unknown",
        "operationId": "unknown_v1_runs__run_id__reservations__reservation_id__unknown_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/reservations/{reservation_id}/release": {
      "post": {
        "summary": "Release",
        "operationId": "release_v1_runs__run_id__reservations__reservation_id__release_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reservation Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_v1_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "after_version",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": -1,
              "default": -1,
              "title": "After Version"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  },
                  "title": "Response Events V1 Runs  Run Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}/baseline": {
      "post": {
        "summary": "Baseline",
        "operationId": "baseline_v1_runs__run_id__baseline_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaselineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Event": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "reservation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reservation Id"
          },
          "version": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Version"
          },
          "kind": {
            "type": "string",
            "enum": [
              "created",
              "reserved",
              "dispatched",
              "unknown",
              "settled",
              "released"
            ],
            "title": "Kind"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "version",
          "kind"
        ],
        "title": "Event"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModelResult": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "usage"
        ],
        "title": "ModelResult"
      },
...
      "Reservation": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "spec": {
            "$ref": "#/components/schemas/ReservationSpec"
          },
          "state": {
            "$ref": "#/components/schemas/ReservationState",
            "default": "reserved"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ]
          },
          "estimate_exceeded": {
            "type": "boolean",
            "title": "Estimate Exceeded",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "spec"
        ],
        "title": "Reservation"
      },
      "ReservationSpec": {
        "properties": {
          "reservation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reservation Id"
          },
          "request_digest": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Request Digest"
          },
          "components": {
            "$ref": "#/components/schemas/Components"
          },
          "max_output_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Max Output Tokens"
          },
          "safety_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Safety Tokens"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "baseline",
              "controller",
              "summarizer",
              "agent",
              "verifier"
            ],
            "title": "Purpose"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reservation_id",
          "request_digest",
          "components",
          "max_output_tokens",
          "safety_tokens",
          "purpose"
        ],
        "title": "ReservationSpec"
      },
      "ReservationState": {
        "type": "string",
        "enum": [
          "reserved",
          "dispatched",
          "unknown",
          "settled",
          "released"
        ],
        "title": "ReservationState"
      },
      "Run": {
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/RunSpec-Output"
          },
          "spent_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Spent Tokens",
            "default": 0
          },
          "reserved_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Reserved Tokens",
            "default": 0
          },
          "input_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Output Tokens",
            "default": 0
          },
          "cached_input_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Cached Input Tokens",
            "default": 0
          },
          "reasoning_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Reasoning Tokens",
            "default": 0
          },
          "cost": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "breached": {
            "type": "boolean",
            "title": "Breached",
            "default": false
          },
          "version": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Version",
            "default": 0
          },
          "available_tokens": {
            "type": "integer",
            "title": "Available Tokens",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spec",
          "available_tokens"
        ],
        "title": "Run"
      },
      "RunSpec-Input": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "budget_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Budget Tokens"
          },
          "context_window": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Context Window"
          },
          "model_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Id"
          },
          "policy_version": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Policy Version"
          },
          "rates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Rates-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "budget_tokens",
          "context_window",
          "model_id",
          "policy_version"
        ],
        "title": "RunSpec"
      },
...
      "Usage": {
        "properties": {
          "input_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Input Tokens"
          },
          "output_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Output Tokens"
          },
          "cached_input_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Cached Input Tokens",
            "default": 0
          },
          "reasoning_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Reasoning Tokens",
            "default": 0
          },
          "source": {
            "type": "string",
            "enum": [
              "provider",
              "simulated"
            ],
            "title": "Source",
            "default": "provider"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "input_tokens",
          "output_tokens"
        ],
        "title": "Usage"
      },
...
    }
  }
}